+++ keras_core/initializers/random_initializers_test.py
@@ -142,37 +142,3 @@ class InitializersTest(testing.TestCase):
 
         with self.assertRaises(ValueError):
             initializers.get("typo")
-
-    def test_zeros_initializer(self):
-        shape = (3, 3)
-
-        initializer = initializers.Zeros()
-        values = initializer(shape=shape)
-        self.assertEqual(values.shape, shape)
-        np_values = backend.convert_to_numpy(values).data
-        self.assertEqual(np_values, np.zeros(shape=shape))
-
-        self.run_class_serialization_test(initializer)
-
-    def test_ones_initializer(self):
-        shape = (3, 3)
-
-        initializer = initializers.Ones()
-        values = initializer(shape=shape)
-        self.assertEqual(values.shape, shape)
-        np_values = backend.convert_to_numpy(values).data
-        self.assertEqual(np_values, np.ones(shape=shape))
-
-        self.run_class_serialization_test(initializer)
-
-    def test_constant_initializer(self):
-        shape = (3, 3)
-        constant_value = 6.
-
-        initializer = initializers.Constant(value=constant_value)
-        values = initializer(shape=shape)
-        self.assertEqual(values.shape, shape)
-        np_values = backend.convert_to_numpy(values).data
-        self.assertEqual(np_values, np.full(shape=shape, fill_value=constant_value))
-
-        self.run_class_serialization_test(initializer)
